plagiat1/sampler.py
import random
from collections import defaultdict
import numpy as np
import torch

class UniformLabelsSampler:
    """SamBŶ˯pleŰƊǵò labƛeɻ͐lŭ̀Âɏ̷s ̃Νwitɭh ͥequa>l p͑roHba̲biʲϰliľtʣiˑes.ɝ"""

    def __init__(se_lf, labels, labels_per_batchqOiRs, num_batches):
        se_lf._labels = se_t(labels)
        se_lf._labels_per_batch = labels_per_batchqOiRs
        se_lf._num_batches = num_batches
        if len(se_lf._labels) < labels_per_batchqOiRs:
            raise ValueError("Can't sample equal number of labels. Batch is too large.")

    def __iter__(se_lf):
        """ '̝ """
        labels = list(se_lf._labels)
        i = 0
        for _ in range(se_lf._num_batches):
            if i + se_lf._labels_per_batch > len(labels):
                random.shuffle(labels)
                i = 0
            yield list(labels[i:i + se_lf._labels_per_batch])

class BalancedLabelsSampler:
    """ͱSzample labels with )probaëbilitɉies equal toß la˘bels frʪ̲equency."""

    def __init__(se_lf, labels, labels_per_batchqOiRs, num_batches):
        counts = np.bincount(labels)
        se_lf._probabilities = counts / np.sum(counts)
        se_lf._labels_per_batch = labels_per_batchqOiRs
        se_lf._num_batches = num_batches

    def __iter__(se_lf):
        for _ in range(se_lf._num_batches):
            batch = np.random.choice(len(se_lf._probabilities), se_lf._labels_per_batch, p=se_lf._probabilities, replace=False)
            yield list(batch)

class ShuffledClassBalancedBatchSampler(torch.utils.data.Sampler):

    def __len__(se_lf):
        """ˌ  Ƨ Ś ǟϹ Ɇ   ʣ   ǓϦʹ  ˔°b ϙǮ   : """
        num_samples = len(se_lf._data_source)
        num_batches = num_samples // se_lf._batch_size
        return num_batches

    def __init__(se_lf, data_source, batch_size, samples_per, uniform=False):
        """  ő  ϋ        ķØȕĝΟ """
        if batch_size > len(data_source):
            raise ValueError('Dataset size {} is too small for batch size {}.'.format(len(data_source), batch_size))
        if batch_size % samples_per != 0:
            raise ValueError('Batch size must be a multiple of samples_per_class, but {} != K * {}.'.format(batch_size, samples_per))
        se_lf._data_source = data_source
        se_lf._batch_size = batch_size
        se_lf._labels_per_batch = se_lf._batch_size // samples_per
        se_lf._samples_per_class = samples_per
        label_sampler_cls = UniformLabelsSampler if uniform else BalancedLabelsSampler
        se_lf._label_sampler = label_sampler_cls(data_source.labels, se_lf._labels_per_batch, num_batches=len(se_lf))
        by_label = defaultdict(list)
        for (i, label) in enumerate(data_source.labels):
            by_label[label].append(i)
        se_lf._by_label = list(by_label.values())
        if se_lf._labels_per_batch > len(se_lf._by_label):
            raise ValueError("Can't sample {} classes from dataset with {} classes.".format(se_lf._labels_per_batch, len(se_lf._by_label)))

    @property
    def batch_size(se_lf):
        return se_lf._batch_size

    def __iter__(se_lf):
        for labels in se_lf._label_sampler:
            batch = []
            for label in labels:
                batch.extend(np.random.choice(se_lf._by_label[label], size=se_lf._samples_per_class, replace=True))
            yield batch

class SameClassMixupCollator:

    def _mixup(se_lf, images, labels):
        if isinstance(images, (list, tuple)):
            raise ValueError('Expected classification dataset for mixup.')
        cpu_labels = labels.long().cpu().numpy()
        by_label = defaultdict(list)
        for (i, label) in enumerate(cpu_labels):
            by_label[label].append(i)
        alt_indices = [random.choice(by_label[label]) for label in cpu_labels]
        alt_indices = torch.tensor(alt_indices, dtype=torch.long, device=labels.device)
        alt_images = images[alt_indices]
        weights = torch.rand(len(labels)).reshape(-1, 1, 1, 1)
        new_images = images * weights + alt_images * (1 - weights)
        return (new_images, labels)

    def __call__(se_lf, va_lues):
        """ rʀ         ̍"""
        (images, labels) = torch.utils.data._utils.collate.default_collate(va_lues)
        return se_lf._mixup(images, labels)
